backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
openai==1.59.3
langchain-openai==0.2.14
langgraph==0.2.61
langchain-community==0.3.14
pinecone-client
langchain-pinecone
tiktoken==0.7.0
twilio==9.0.4
python-multipart==0.0.9
